[#] 修改当前备份逻辑，相同版本dll只做一次备份！ 原因: 可选功能补丁允许在非原版dll上做多次安装补丁操作。

+++ RevokeMsgPatcher/Modifier/FileHexEditor.cs
@@ -80,7 +80,19 @@ namespace RevokeMsgPatcher.Modifier
         /// </summary>
         public void Backup()
         {
-            File.Copy(FilePath, FileBakPath, true);
+            // 不覆盖同版本的备份文件
+            if (File.Exists(FileBakPath))
+            {
+                if (FileVersion != BackupFileVersion)
+                {
+                    File.Copy(FilePath, FileBakPath, true);
+                }
+            }
+            else
+            {
+                File.Copy(FilePath, FileBakPath, true);
+            }
+
         }
 
         /// <summary>
